Places validation
Made it so the validation section of the code would check for a
character sequence defined to eventLocation. This stores the address of
the event. Later down the line we could add the name to it, but for now
I just wanted to get something from the place object to store their.